+++ cucumber-tsflow/src/api/convert-configuration.ts
@@ -8,6 +8,7 @@ const logger = createLogger('convert');
 
 export interface IConfigurationExt extends IConfiguration {
 	experimentalDecorators: boolean;
+	parallelLoad: boolean | number;
 }
 
 export async function convertConfiguration(
@@ -43,6 +44,7 @@ export async function convertConfiguration(
 
 		logger.checkpoint('Building runtime config', {
 			parallel: flatConfiguration.parallel,
+			parallelLoad: flatConfiguration.parallelLoad,
 			experimentalDecorators: flatConfiguration.experimentalDecorators
 		});
 		const runtime = {
@@ -51,6 +53,7 @@ export async function convertConfiguration(
 			failFast: flatConfiguration.failFast,
 			filterStacktraces: !flatConfiguration.backtrace,
 			parallel: flatConfiguration.parallel,
+			parallelLoad: flatConfiguration.parallelLoad,
 			retry: flatConfiguration.retry,
 			retryTagFilter: flatConfiguration.retryTagFilter,
 			strict: flatConfiguration.strict,