+++ cucumber-tsflow/src/runtime/types.ts
@@ -12,6 +12,7 @@ export interface IMessageData {
 
 export interface ITsFlowRunOptionsRuntime extends IRunOptionsRuntime {
 	experimentalDecorators: boolean;
+	parallelLoad: boolean | number;
 }
 export interface ITsFlowRunConfiguration extends IRunConfiguration {
 	runtime: ITsFlowRunOptionsRuntime;
@@ -19,6 +20,7 @@ export interface ITsFlowRunConfiguration extends IRunConfiguration {
 
 export interface TsFlowRuntimeOptions extends RuntimeOptions {
 	experimentalDecorators: boolean;
+	parallelLoad: boolean | number;
 }
 
 export interface InitializeTsflowCommand extends InitializeCommand {